Version 2.5.0

+++ sdk/appcenter-crashes/src/main/java/com/microsoft/appcenter/crashes/Crashes.java
@@ -29,6 +29,7 @@ import com.microsoft.appcenter.crashes.model.ErrorReport;
 import com.microsoft.appcenter.crashes.model.NativeException;
 import com.microsoft.appcenter.crashes.model.TestCrashException;
 import com.microsoft.appcenter.crashes.utils.ErrorLogHelper;
+import com.microsoft.appcenter.ingestion.models.Device;
 import com.microsoft.appcenter.ingestion.models.Log;
 import com.microsoft.appcenter.ingestion.models.json.DefaultLogSerializer;
 import com.microsoft.appcenter.ingestion.models.json.LogFactory;
@@ -159,6 +160,11 @@ public class Crashes extends AbstractAppCenterService {
     private long mInitializeTimestamp;
 
     /**
+     * Cached device info.
+     */
+    private Device mDevice;
+
+    /**
      * Crash handler.
      */
     private UncaughtExceptionHandler mUncaughtExceptionHandler;
@@ -245,31 +251,29 @@ public class Crashes extends AbstractAppCenterService {
     }
 
     /**
-     * Track an exception.
-     * TODO the backend does not support that service yet, will be public method later.
+     * Track a handled error.
      *
-     * @param throwable An exception.
+     * @param throwable The throwable describing the handled error.
      */
-    @SuppressWarnings("SameParameterValue")
-    static void trackException(@NonNull Throwable throwable) {
-        trackException(throwable, null);
+    @SuppressWarnings({"SameParameterValue", "WeakerAccess"})
+    public static void trackError(Throwable throwable) {
+        trackError(throwable, null, null);
     }
 
     /**
-     * Track a custom exception with name and optional properties.
+     * Track a handled error with name and optional properties and attachments.
      * The name parameter can not be null or empty. Maximum allowed length = 256.
      * The properties parameter maximum item count = 5.
      * The properties keys can not be null or empty, maximum allowed key length = 64.
      * The properties values can not be null, maximum allowed value length = 64.
      * Any length of name/keys/values that are longer than each limit will be truncated.
-     * TODO the backend does not support that service yet, will be public method later.
      *
-     * @param throwable  An exception.
-     * @param properties Optional properties.
+     * @param throwable   The throwable describing the handled error.
+     * @param properties  Optional properties.
+     * @param attachments Optional attachments.
      */
-    static void trackException(@NonNull Throwable throwable, Map<String, String> properties) {
-        Map<String, String> validatedProperties = ErrorLogHelper.validateProperties(properties, "HandledError");
-        getInstance().queueException(throwable, validatedProperties);
+    public static void trackError(Throwable throwable, Map<String, String> properties, Iterable<ErrorAttachmentLog> attachments) {
+        getInstance().queueException(throwable, properties, attachments);
     }
 
     /**
@@ -417,7 +421,7 @@ public class Crashes extends AbstractAppCenterService {
                 }
 
                 @Override
-                public void onConfigurationChanged(Configuration newConfig) {
+                public void onConfigurationChanged(@NonNull Configuration newConfig) {
                 }
 
                 @Override
@@ -429,10 +433,13 @@ public class Crashes extends AbstractAppCenterService {
         } else {
 
             /* Delete all files. */
-            for (File file : ErrorLogHelper.getErrorStorageDirectory().listFiles()) {
-                AppCenterLog.debug(LOG_TAG, "Deleting file " + file);
-                if (!file.delete()) {
-                    AppCenterLog.warn(LOG_TAG, "Failed to delete file " + file);
+            File[] files = ErrorLogHelper.getErrorStorageDirectory().listFiles();
+            if (files != null) {
+                for (File file : files) {
+                    AppCenterLog.debug(LOG_TAG, "Deleting file " + file);
+                    if (!file.delete()) {
+                        AppCenterLog.warn(LOG_TAG, "Failed to delete file " + file);
+                    }
                 }
             }
             AppCenterLog.info(LOG_TAG, "Deleted crashes local files");
@@ -569,12 +576,18 @@ public class Crashes extends AbstractAppCenterService {
         };
     }
 
+    synchronized Device getDeviceInfo(Context context) throws DeviceInfoHelper.DeviceInfoException {
+        if (mDevice == null) {
+            mDevice = DeviceInfoHelper.getDeviceInfo(context);
+        }
+        return mDevice;
+    }
+
     /**
      * Get initialization timestamp.
      *
      * @return initialization timestamp expressed using {@link System#currentTimeMillis()}.
      */
-    @VisibleForTesting
     synchronized long getInitializeTimestamp() {
         return mInitializeTimestamp;
     }
@@ -582,17 +595,18 @@ public class Crashes extends AbstractAppCenterService {
     /**
      * Send an handled exception.
      *
-     * @param throwable  An handled exception.
-     * @param properties optional properties.
+     * @param throwable   An handled exception.
+     * @param properties  optional properties.
+     * @param attachments optional attachments.
      */
-    private synchronized void queueException(@NonNull final Throwable throwable, Map<String, String> properties) {
+    private synchronized void queueException(@NonNull final Throwable throwable, Map<String, String> properties, Iterable<ErrorAttachmentLog> attachments) {
         queueException(new ExceptionModelBuilder() {
 
             @Override
             public Exception buildExceptionModel() {
                 return ErrorLogHelper.getModelExceptionFromThrowable(throwable);
             }
-        }, properties);
+        }, properties, attachments);
     }
 
     /**
@@ -600,30 +614,43 @@ public class Crashes extends AbstractAppCenterService {
      *
      * @param modelException An handled exception already in JSON model form.
      * @param properties     optional properties.
+     * @param attachments    optional attachments.
+     * @return handled error ID.
      */
-    synchronized void queueException(@NonNull final Exception modelException, Map<String, String> properties) {
-        queueException(new ExceptionModelBuilder() {
+    synchronized UUID queueException(@NonNull final Exception modelException, Map<String, String> properties, Iterable<ErrorAttachmentLog> attachments) {
+        return queueException(new ExceptionModelBuilder() {
 
             @Override
             public Exception buildExceptionModel() {
                 return modelException;
             }
-        }, properties);
+        }, properties, attachments);
     }
 
-    private synchronized void queueException(@NonNull final ExceptionModelBuilder exceptionModelBuilder, final Map<String, String> properties) {
+    private synchronized UUID queueException(@NonNull final ExceptionModelBuilder exceptionModelBuilder, Map<String, String> properties, final Iterable<ErrorAttachmentLog> attachments) {
+
+        /* Snapshot userId as early as possible. */
+        final String userId = UserIdContext.getInstance().getUserId();
+        final UUID errorId = UUID.randomUUID();
+        final Map<String, String> validatedProperties = ErrorLogHelper.validateProperties(properties, "HandledError");
         post(new Runnable() {
 
             @Override
             public void run() {
+
+                /* First send the handled error. */
                 HandledErrorLog errorLog = new HandledErrorLog();
-                errorLog.setId(UUID.randomUUID());
-                errorLog.setUserId(UserIdContext.getInstance().getUserId());
+                errorLog.setId(errorId);
+                errorLog.setUserId(userId);
                 errorLog.setException(exceptionModelBuilder.buildExceptionModel());
-                errorLog.setProperties(properties);
+                errorLog.setProperties(validatedProperties);
                 mChannel.enqueue(errorLog, ERROR_GROUP, Flags.DEFAULTS);
+
+                /* Then attachments if any. */
+                sendErrorAttachment(errorId, attachments);
             }
         });
+        return errorId;
     }
 
     private void initialize() {
@@ -695,7 +722,7 @@ public class Crashes extends AbstractAppCenterService {
              */
             errorLog.setUserId(UserIdContext.getInstance().getUserId());
             try {
-                errorLog.setDevice(DeviceInfoHelper.getDeviceInfo(mContext));
+                errorLog.setDevice(getDeviceInfo(mContext));
                 errorLog.getDevice().setWrapperSdkName(WRAPPER_SDK_NAME_NDK);
                 saveErrorLogFiles(new NativeException(), errorLog);
                 if (!logFile.renameTo(dest)) {
@@ -765,7 +792,7 @@ public class Crashes extends AbstractAppCenterService {
             }
         }
         mHasReceivedMemoryWarningInLastSession = isMemoryRunningLevelWasReceived(SharedPreferencesManager.getInt(PREF_KEY_MEMORY_RUNNING_LEVEL, -1));
-        if(mHasReceivedMemoryWarningInLastSession) {
+        if (mHasReceivedMemoryWarningInLastSession) {
             AppCenterLog.debug(LOG_TAG, "The application received a low memory warning in the last session.");
         }
         SharedPreferencesManager.remove(PREF_KEY_MEMORY_RUNNING_LEVEL);
@@ -979,9 +1006,10 @@ public class Crashes extends AbstractAppCenterService {
     /**
      * Send error attachment logs through channel.
      */
+    @WorkerThread
     private void sendErrorAttachment(UUID errorId, Iterable<ErrorAttachmentLog> attachments) {
         if (attachments == null) {
-            AppCenterLog.debug(LOG_TAG, "CrashesListener.getErrorAttachments returned null, no additional information will be attached to log: " + errorId.toString());
+            AppCenterLog.debug(LOG_TAG, "Error report: " + errorId.toString() + " does not have any attachment.");
         } else {
             int totalErrorAttachments = 0;
             for (ErrorAttachmentLog attachment : attachments) {
@@ -995,7 +1023,7 @@ public class Crashes extends AbstractAppCenterService {
                         AppCenterLog.error(LOG_TAG, "Not all required fields are present in ErrorAttachmentLog.");
                     }
                 } else {
-                    AppCenterLog.warn(LOG_TAG, "Skipping null ErrorAttachmentLog in CrashesListener.getErrorAttachments.");
+                    AppCenterLog.warn(LOG_TAG, "Skipping null ErrorAttachmentLog.");
                 }
             }
             if (totalErrorAttachments > MAX_ATTACHMENT_PER_CRASH) {
@@ -1155,6 +1183,7 @@ public class Crashes extends AbstractAppCenterService {
     /**
      * Implementation of {@link WrapperSdkExceptionManager#sendErrorAttachments(String, Iterable)}.
      */
+    @WorkerThread
     void sendErrorAttachments(final String errorReportId, final Iterable<ErrorAttachmentLog> attachments) {
         post(new Runnable() {
